requirements.txt
requests==2.32.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
python-dotenv==1.0.1
